Add Nominatim API and more responsive design
Nominatim API allows knowing an address by its coordinates. This way,
if someone clicks on the map, I can give the city, state and street
of the location. And that is exactly what I have done (without the
street part yet).
I also used a media query in the CSS of the registration page which
made the site looks way better in more narrow screens.